Before dev

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.skypro.homework.entity.AdEntity;
 @Repository
-public interface AdRepository extends JpaRepository<AdEntity, Integer> {
+public interface AdRepository extends JpaRepository<AdEntity, Long> {
 
 }
